requirements.txt
fastapi
uvicorn
boto3
python-pptx
python-dotenv
openai
